requirements.txt
pyarrow
pandas
ib_insync
yfinance
pyodbc
python-dotenv
sqlalchemy